Refactor to break up runner methods

+++ lib/gitdocs.rb
@@ -31,8 +31,10 @@ module Gitdocs
         t
       end
       sleep 1
-      pid = fork { Server.new(*runners).start }
-      at_exit { Process.kill("KILL", pid) rescue nil }
+      unless defined?(pid) && pid
+        pid = fork { Server.new(*runners).start }
+        at_exit { Process.kill("KILL", pid) rescue nil }
+      end
       puts "Watch threads: #{threads.map { |t| "Thread status: '#{t.status}', running: #{t.alive?}" }}" if debug
       puts "Joined #{threads.size} watch threads...running" if debug
       threads.each(&:join)
